+++ src/main/java/org/swisspush/redisques/util/RedisquesAPI.java
@@ -55,9 +55,11 @@ public class RedisquesAPI {
     public static final String PER_QUEUE_CONFIG_ENQUEUE_MAX_DELAY_MILLIS =  "enqueueMaxDelayMillis";
     public static final String PER_QUEUE_CONFIG_MAX_QUEUE_ENTRIES =  "maxQueueEntries";
     public static final String PER_QUEUE_CONFIG_MAX_QUEUE_PATROL_LIMIT =  "queuePatrolLimit";
+    public static final String PER_QUEUE_CONFIG_MAXIMUM_ITEM_IN_BATCH_DISPATCH =  "maximumItemInBatchDispatch";
+    public static final String PER_QUEUE_CONFIG_MINIMUM_ITEM_IN_BATCH_DISPATCH =  "minimumItemInBatchDispatch";
+    public static final String PER_QUEUE_CONFIG_MAX_BATCH_DISPATCH_WAIT_TIMEOUT =  "maxBatchItemDispatchWaitTimeout";
     public static final String PER_QUEUE_CONFIG_NAME =  "configName";
     public static final String PER_QUEUE_CONFIG_PATTERN =  "pattern";
-
 
     private static final Logger log = LoggerFactory.getLogger(RedisquesAPI.class);